+++ packages/mg-components/src/components/molecules/inputs/mg-input-toggle/test/mg-input-toggle.e2e.ts
@@ -1,4 +1,4 @@
-import { renderAttributes, renderProperties } from '../../../../../utils/e2e.test.utils';
+import { renderAttributes, renderProperties } from '@mgdis/playwright-helpers';
 import { setPageContent, describe, testEach, expect, PageType, test } from '../../../../../utils/playwright.e2e.test.utils';
 import { MgInputToggle } from '../mg-input-toggle';