+++ packages/stencil-helpers/src/storybook/stencil-shim.d.ts
@@ -1,0 +1,1 @@
+declare module '@stencil/core/internal/client';